chore: rename the jira-lint workflow to distinguish from shared workflows
This repository will contain a collection of workflows that are intended
to be shared within Invitation Homes.  In addition, some of the workflows
exist to check code quality and that conventions are adhered to when commiting
to this repository.  To distinguish between the 2, any workflows that are prefix
with __ are not intended to be shared but are executed as contributions are
made to this repository.

jira: ENG-237